Rename Capability.* to PeerCapability.*
