+++ test/unittests/libp2p/capability.cpp
@@ -25,7 +25,7 @@ along with cpp-ethereum.  If not, see <http://www.gnu.org/licenses/>.
 #include <libp2p/Common.h>
 #include <libp2p/Host.h>
 #include <libp2p/Session.h>
-#include <libp2p/Capability.h>
+#include <libp2p/PeerCapability.h>
 #include <libp2p/HostCapability.h>
 #include <test/tools/libtesteth/TestOutputHelper.h>
 #include <test/tools/libtesteth/Options.h>
